requirements.txt
asgiref==3.8.1
certifi==2025.1.31
Django==4.2.19
django-cors-headers==4.7.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
elastic-transport==8.17.1
elasticsearch==8.17.2
PyJWT==2.9.0
rest-framework-simplejwt==0.0.2
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
